+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [1.5.7-SNAPSHOT]
 
+- Node.toggleClass(class, observableBooleanValue)
+- Removed cell as `this` for `cellCache`. The cell could change, so taking it into account was a mistake.
 - App MainView parameter can now be a `Fragment` as well as `View`
 - ListView `cellCache` provider to create a cached graphic node per item
 - Kotlin 1.0.4